services/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
requests
requests-toolbelt
passlib
python-multipart
python-jose[cryptography]
pydantic-settings
email-validator
pillow
reportlab
celery[redis]
redis